backend/requirements.txt
langgraph
langchain
python-dotenv
langchain-google-genai
google-generativeai
beautifulsoup4
pymupdf
